+++ data/shaders/env/skydome.f
@@ -9,7 +9,7 @@ in vec2 f_texcoord;
 layout (location = 0) out vec4 out_color;
 layout (location = 1) out vec4 out_pos;
 layout (location = 2) out vec4 out_norm;
-layout (location = 3) out float out_coverage;
+layout (location = 3) out int out_coverage;
 
 void main() {
 
@@ -18,5 +18,5 @@ void main() {
 	out_color = texture(tex, sky_uv);
 	out_pos = vec4(0.0f);
 	out_norm = vec4(0.0f, 0.0f, 0.0f, 1.0f);
-	out_coverage = 0.0f;
+	out_coverage = 0;
 }
